connectors/google_cloud_storage/requirements.txt
aiofiles==24.1.0
certifi==2025.8.3
colorlog==6.9.0
fastapi==0.124.0
httpx==0.28.1
pydantic==2.11.7
pydantic_settings==2.10.1
Requests==2.32.4
starlette==0.49.1
tenacity==9.1.2
uvicorn==0.35.0
aiohttp>=3.8.0
google-cloud-storage>=3.3.0
google-cloud-pubsub==2.33.0
typer==0.16.0